refactor: EVM listener and intent processing
Updates the EVM listener to use `watchContractEvent` with ABI from `@eco-foundation/routes-ts` for more efficient intent processing.

Renames `intentId` to `intentHash` for clarity and consistency.

Submits intent to fulfillment service on event, removing the need for polling.

Removes strategy determination from `submitIntent` and relies on the listener to pass the desired strategy to fulfillment service.