feat: fetches prover addresses from Eco package
Updates the configuration to fetch prover addresses
(HyperProver, MetaProver) from the `@eco-foundation/routes-ts` package.

This ensures the application uses the latest and most accurate
prover addresses for interacting with the Eco protocol, retrieved
from the Eco NPM package.

Additionally, the ProverType enum is updated to a const object, allowing
for value extraction for runtime type checking.