feat: update Eco Adapter and Arbiter contracts
Introduces the Eco Adapter and Eco Arbiter contracts, defining their interfaces,
functions, events, and error handling. These contracts are crucial for
facilitating cross-chain transactions and claim verification within the Rhinestone
protocol. The contracts are added as ABI definitions in TypeScript.

Also, refactors the validator to allow more flexibility in the adapter validation
to avoid unnecessary errors. It also removes the claimHashOracle requirement
to allow more flexibility for the intent fulfillment.